Remove deprecated ZoeDepth data, model, and trainer modules
