--- src/zensvi/cv/depth_estimation/zoedepth/utils/geometry.py
@@ -1,108 +1,0 @@
-# MIT License
-
-# Copyright (c) 2022 Intelligent Systems Lab Org
-
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-# File author: Shariq Farooq Bhat
-
-import numpy as np
-
-
-def get_intrinsics(H, W):
-    """Calculate the intrinsic matrix for a pinhole camera model.
-
-    This function assumes a field of view of 55 degrees and a central principal point.
-
-    Args:
-        H (int): Height of the image.
-        W (int): Width of the image.
-
-    Returns:
-        np.ndarray: A 3x3 intrinsic matrix.
-    """
-    f = 0.5 * W / np.tan(0.5 * 55 * np.pi / 180.0)
-    cx = 0.5 * W
-    cy = 0.5 * H
-    return np.array([[f, 0, cx], [0, f, cy], [0, 0, 1]])
-
-
-def depth_to_points(depth, R=None, t=None):
-    """Convert depth map to 3D points.
-
-    Args:
-        depth (np.ndarray): Depth map of shape (B, H, W).
-        R (np.ndarray, optional): Rotation matrix of shape (3, 3). Defaults to identity matrix.
-        t (np.ndarray, optional): Translation vector of shape (3,). Defaults to zero vector.
-
-    Returns:
-        np.ndarray: 3D points of shape (B, 3).
-    """
-    K = get_intrinsics(depth.shape[1], depth.shape[2])
-    Kinv = np.linalg.inv(K)
-    if R is None:
-        R = np.eye(3)
-    if t is None:
-        t = np.zeros(3)
-
-    # M converts from your coordinate to PyTorch3D's coordinate system
-    M = np.eye(3)
-    M[0, 0] = -1.0
-    M[1, 1] = -1.0
-
-    height, width = depth.shape[1:3]
-
-    x = np.arange(width)
-    y = np.arange(height)
-    coord = np.stack(np.meshgrid(x, y), -1)
-    coord = np.concatenate((coord, np.ones_like(coord)[:, :, [0]]), -1)  # z=1
-    coord = coord.astype(np.float32)
-    coord = coord[None]  # bs, h, w, 3
-
-    D = depth[:, :, :, None, None]
-    pts3D_1 = D * Kinv[None, None, None, ...] @ coord[:, :, :, :, None]
-    pts3D_1 = M[None, None, None, ...] @ pts3D_1
-    pts3D_2 = R[None, None, None, ...] @ pts3D_1 + t[None, None, None, :, None]
-    return pts3D_2[:, :, :, :3, 0][0]
-
-
-def create_triangles(h, w, mask=None):
-    """Create mesh triangle indices from a given pixel grid size.
-
-    This function is not and need not be differentiable as triangle indices are fixed.
-
-    Args:
-        h (int): Height of the image.
-        w (int): Width of the image.
-        mask (np.ndarray, optional): A mask to filter triangles. Defaults to None.
-
-    Returns:
-        np.ndarray: A 2D array of triangle indices with shape (2(W-1)(H-1), 3).
-    """
-    x, y = np.meshgrid(range(w - 1), range(h - 1))
-    tl = y * w + x
-    tr = y * w + x + 1
-    bl = (y + 1) * w + x
-    br = (y + 1) * w + x + 1
-    triangles = np.array([tl, bl, tr, br, tr, bl])
-    triangles = np.transpose(triangles, (1, 2, 0)).reshape(((w - 1) * (h - 1) * 2, 3))
-    if mask is not None:
-        mask = mask.reshape(-1)
-        triangles = triangles[mask[triangles].all(1)]
-    return triangles
